Create necessary files for workshops page to work
