+++ app/Http/Controllers/HomePageController.php
@@ -2,62 +2,62 @@
 
 namespace App\Http\Controllers;
 
-use App\Services\HomePageService;
+use App\Interfaces\HomePageRepositoryInterface;
 use Illuminate\Database\Eloquent\Collection;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 
 class HomePageController extends Controller
 {
-    protected HomePageService $homePageService;
+    protected HomePageRepositoryInterface $homePageRepository;
 
-    public function __construct(HomePageService $homePageService)
+    public function __construct(HomePageRepositoryInterface $homePageRepository)
     {
-        $this->homePageService = $homePageService;
+        $this->homePageRepository = $homePageRepository;
     }
 
     public function getCarouselImages(): Collection
     {
-        return $this->homePageService->getCarouselImages();
+        return $this->homePageRepository->getCarouselImages();
     }
 
     public function storeCarouselImage(Request $request): void
     {
-        $this->homePageService->storeCarouselImage($request);
+        $this->homePageRepository->storeCarouselImage($request);
     }
 
     public function deleteCarouselImage(int $id): void
     {
-        $this->homePageService->deleteCarouselImage($id);
+        $this->homePageRepository->deleteCarouselImage($id);
     }
 
     public function getPosts(): Collection
     {
-        return $this->homePageService->getPosts();
+        return $this->homePageRepository->getPosts();
     }
 
     public function storePost(Request $request): void
     {
-        $this->homePageService->storePost($request);
+        $this->homePageRepository->storePost($request);
     }
 
     public function deletePost(int $id): void
     {
-        $this->homePageService->deletePost($id);
+        $this->homePageRepository->deletePost($id);
     }
 
     public function getMetaData(): Response
     {
-        return $this->homePageService->getMetaData();
+        return $this->homePageRepository->getMetaData();
     }
 
     public function storeMetaData(Request $request): void
     {
-        $this->homePageService->storeMetaData($request);
+        $this->homePageRepository->storeMetaData($request);
     }
 
     public function deleteMetaData(int $id): void
     {
-        $this->homePageService->deleteMetaData($id);
+        $this->homePageRepository->deleteMetaData($id);
     }
 }
